Fix flaky Figwheel behavior
It turns out that Figwheel, when launched via Figwheel Sidecar (but not
when launched via lein-figwheel), does not serve files from resources/
unless the directory exists *before* starting Figwheel.

Checking resources/ into version control and removing it from
:clean-targets should fix this flaky behavior. An ideal solution would
be to make Figwheel work even if resources/ does not exist when it is
launched, but in the meantime this fix will ensure correct behavior.

Since previous attempts to fix the problem (by adding :http-server-root
and downgrading figwheel-sidecar) were misguided and introduced unneeded
complexity, they are reverted in this commit.

Also, the README was updated to reflect the dependency upgrades.